requirements.txt
discord.py
openai
requests
beautifulsoup4
duckduckgo-search